Major refactor on the tween structs
