+++ Config/DefaultNsTween.ini
@@ -1,0 +1,2 @@
+﻿[CoreRedirects]
++ClassRedirects=(OldName="/Script/NsTween.NsTweenBlueprintLibrary",NewName="/Script/NsTween.NsTweenFunctionLibrary")
