+++ Source/NsTween/Public/Classes/NsTweenCore.h
@@ -2,8 +2,7 @@
 
 #pragma once
 
-#include "NsTweenEasing.h"
-#include "NsTweenInstance.h"
+#include "Library/NsTweenTypeLibrary.h"
 #include "NsTweenManager.h"
 
 NSTWEEN_API DECLARE_LOG_CATEGORY_EXTERN(LogNsTween, Log, All)
@@ -23,10 +22,10 @@ public:
     static void Deinitialize();
 
     /** Ensure there are at least this many tweens in the recycle pool. Call this at game startup to increase your initial capacity for each type of tween, if you know you will be needing more and don't want to allocate memory during the game */
-    static void EnsureCapacity(int NumFloatTweens, int NumVectorTweens, int NumVector2DTweens, int NumQuatTweens);
+    static void EnsureCapacity(const int32 NumFloatTweens, const int32 NumVectorTweens, const int32 NumVector2DTweens, const int32 NumQuatTweens);
 
     /** Add more tweens to the recycle pool. Call this at game startup to increase your initial capacity if you know you will be needing more and don't want to allocate memory during the game */
-    static void EnsureCapacity(int NumTweens);
+    static void EnsureCapacity(const int32 NumTweens);
 
     /** Update */
     static void Update(float UnscaledDeltaSeconds, float DilatedDeltaSeconds, bool bIsGamePaused);
@@ -41,16 +40,16 @@ public:
     static float Ease(float T, ENsTweenEase EaseType);
 
     /** Play Tween - Float */
-    static NsTweenInstanceFloat& Play(float Start, float End, float DurationSecs, ENsTweenEase EaseType, TFunction<void(float)> OnUpdate);
+    static FNsTweenInstanceFloat& Play(float Start, float End, float DurationSecs, ENsTweenEase EaseType, TFunction<void(float)> OnUpdate);
 
     /** Play Tween - Vector */
-    static NsTweenInstanceVector& Play(const FVector& Start, const FVector& End, float DurationSecs, ENsTweenEase EaseType, TFunction<void(FVector)> OnUpdate);
+    static FNsTweenInstanceVector& Play(const FVector& Start, const FVector& End, float DurationSecs, ENsTweenEase EaseType, TFunction<void(FVector)> OnUpdate);
 
     /** Play Tween - Vector2D */
-    static NsTweenInstanceVector2D& Play(FVector2D Start, FVector2D End, float DurationSecs, ENsTweenEase EaseType, TFunction<void(FVector2D)> OnUpdate);
+    static FNsTweenInstanceVector2D& Play(FVector2D Start, FVector2D End, float DurationSecs, ENsTweenEase EaseType, TFunction<void(FVector2D)> OnUpdate);
 
     /** Play Tween - Quat */
-    static NsTweenInstanceQuat& Play(const FQuat& Start, const FQuat& End, float DurationSecs, ENsTweenEase EaseType, TFunction<void(FQuat)> OnUpdate);
+    static FNsTweenInstanceQuat& Play(const FQuat& Start, const FQuat& End, float DurationSecs, ENsTweenEase EaseType, TFunction<void(FQuat)> OnUpdate);
 
     /** For Each Manager */
     template <typename Func>
@@ -102,16 +101,16 @@ public:
 private:
 
     /** Float tween manager */
-    static NsTweenManager<NsTweenInstanceFloat>* FloatTweenManager;
+    static NsTweenManager<FNsTweenInstanceFloat>* FloatTweenManager;
 
     /** Vector tween manager */
-    static NsTweenManager<NsTweenInstanceVector>* VectorTweenManager;
+    static NsTweenManager<FNsTweenInstanceVector>* VectorTweenManager;
 
     /** Vector2D tween manager */
-    static NsTweenManager<NsTweenInstanceVector2D>* Vector2DTweenManager;
+    static NsTweenManager<FNsTweenInstanceVector2D>* Vector2DTweenManager;
 
     /** Quat tween manager */
-    static NsTweenManager<NsTweenInstanceQuat>* QuatTweenManager;
+    static NsTweenManager<FNsTweenInstanceQuat>* QuatTweenManager;
 
     /** Reserved floats */
     static int32 CurrentReservedFloat;
